perf(gemma4): stage the paged KV cache in bf16 rather than fp32
torch.zeros(cache_shape) took torch's default float32 while cache_dtype is bfp8
for this model (precision_overrides.json), so every K/V allocation staged twice
the bytes it needed. At a 256k context the shape is [4096, 4, 64, 256] = 1.00 GiB
per tensor, built for every layer, K and V, replicated to all 32 devices -- and
1.00 GiB is also the per-device sysmem window. bf16 halves it.

bf16 is what ttnn wants for a BFP target anyway: from_torch converts
source -> bf16 -> bf8/4 and deliberately avoids float32 as the intermediate
("float32 as an intermediate type is not used due to limited amount of L1
memory"), so the fp32 source only bought an extra conversion on the way in. The
values are zeros, which round-trip exactly in any dtype, so nothing is lost --
including if kv_cache is overridden to bf16 or fp32.

init_kv_staging had the same pattern on a smaller tensor; fixed alongside it.

This is NOT a fix for the intermittent SIGBUS seen during model construction at
256k. That fault reproduces and passes with byte-identical code, so its cause is
still unknown; this only shrinks the largest host->device staging allocation in
the run, which is a plausible lever and a real waste regardless. Unverified
against the fault.

+++ models/demos/gemma4/tt/attention/kv_cache.py
@@ -88,8 +88,16 @@ def init_kv_cache(
     # downstream where SDPA expects the narrower format.
     dtype_suffix = f"_{dtype_to_str(cache_dtype)}"
 
+    # Stage in bf16, not torch's default float32. This tensor is
+    # [max_num_blocks, kv_heads, block_size, head_dim], which reaches 1.00 GiB at a 256k
+    # context -- the same size as the per-device sysmem window -- and it is built for every
+    # layer, K and V, replicated to every device. bf16 halves that. It is also the
+    # intermediate ttnn itself uses for the bfp8/bfp4 targets this cache actually runs at
+    # (from_torch converts source -> bf16 -> bf8/4 and deliberately avoids float32), so the
+    # fp32 source only bought an extra conversion. The values are zeros, which round-trip
+    # exactly in any dtype, so nothing is lost.
     k_cache = ttnn.as_tensor(
-        torch.zeros(cache_shape),
+        torch.zeros(cache_shape, dtype=torch.bfloat16),
         device=mesh_device,
         layout=ttnn.TILE_LAYOUT,
         dtype=cache_dtype,
@@ -99,7 +107,7 @@ def init_kv_cache(
     )
 
     v_cache = ttnn.as_tensor(
-        torch.zeros(cache_shape),
+        torch.zeros(cache_shape, dtype=torch.bfloat16),
         device=mesh_device,
         layout=ttnn.TILE_LAYOUT,
         dtype=cache_dtype,
@@ -142,7 +150,7 @@ def init_kv_staging(
 
     def _zeros():
         return ttnn.as_tensor(
-            torch.zeros(stage_shape),
+            torch.zeros(stage_shape, dtype=torch.bfloat16),
             device=mesh_device,
             layout=ttnn.TILE_LAYOUT,
             dtype=cache_dtype,
